ECE-482 Added video log message for Jenkins
As of now Jenkins doesn't seem to be recording video, or at least not saving it. These debug messages will let us know hopefully which problem it is.
If the messages show up, it's saving the improperly somehow, if they don't show up, then Jenkins isn't doing anything with video.